Unify common browser+node tests
This removes code duplication and makes tests maintenance easier
